Explicit casting from double to int
git-svn-id: https://ferma.svn.sf.net/svnroot/ferma/trunk/FermaNext@198 d51c77f7-620d-0410-a109-9a2f07eaa567

+++ src/TrussUnit/TrussUnitWindow.cpp
@@ -667,8 +667,8 @@ QPoint TrussUnitWindow::getLineSegmentsIntersectionPoint ( QPoint p11, QPoint p1
         // intersection point is within both line segments
         double x = p11.x() + ( p12.x() - p11.x() ) * Ua2;
         double y = p11.y() + ( p12.y() - p11.y() ) * Ua2;
-        crossPoint.setX ( x );
-        crossPoint.setY ( y );
+        crossPoint.setX ( (int)x );
+        crossPoint.setY ( (int)y );
         return crossPoint;
     }
     else
